Refactor message parsing to use proper Pydantic models throughout
This commit completes the refactoring of message parsing to use Pydantic
models instead of dict wrangling. Key changes:

## Translation Module
- Create new translation.py module for format conversion functions
- Move anthropic_messages_to_standard from run_eval.py to translation.py
- Keep anthropic_to_chat_messages and anthropic_to_responses_input in translation

## Type Safety Improvements
1. responses_to_ccr_messages now takes list[ResponseOutputMessage] instead of Any
   - Parse input early using parse_response_messages
   - Add validation checks at call sites

2. Create Grade pydantic model for score + rationale
   - Replace dict[str, Any] with proper Grade type
   - Add Field constraints (score: 1-5, rationale: str)
   - Update EvalSampleRecord.grade to use Grade | None

3. Update parse_grade_from_responses to return Grade instead of dict
   - Use Grade.model_validate_json/model_validate
   - Update all call sites to use grade.score instead of parsed.get("score")

## Code Cleanup
- Inline sys_val variable (use ar.system or "" directly)
- Fix grade_rec dict wrangling to use g_obj.response instead of grade_rec["response"]
- Use g_obj.correlation_id instead of grade_rec.get("correlation_id")
- Fix CCRRequest reference to use Request (correct name from schemas)
- Remove unused imports (extract_text_content from run_eval.py)

All changes ensure Pydantic validation happens early and objects are used
consistently throughout instead of dict access with type: ignore comments.
